Add support for DIMM modules of X299 Chipset (#7). For Windows WinRing0 and PawnIO is supported. Linux is supported too.
Update manufacturer mapping.
Update dependencies.

+++ RAMSPDToolkit/SPD/Interop/DDR4Constants.cs
@@ -33,7 +33,7 @@ namespace RAMSPDToolkit.SPD.Interop
         public const byte SPD_DDR4_THERMAL_SENSOR_MANUFACTURER           = 0x06;
         public const byte SPD_DDR4_THERMAL_SENSOR_DEVICEID               = 0x07;
 
-        public const ushort SPD_DDR4_MODULE_SPD_REVISION           = 0x001;
+        public const byte   SPD_DDR4_MODULE_SPD_REVISION           = 0x001;
         public const byte   SPD_DDR4_MODULE_MEMORY_TYPE            = 0x002;
         public const ushort SPD_DDR4_MODULE_DENSITY_BANKS          = 0x004;
         public const ushort SPD_DDR4_MODULE_PRIMARY_PACKAGE_TYPE   = 0x006;
